+++ minix/lib/libvtreefs/file.c
@@ -90,7 +90,7 @@ fs_read(ino_t ino_nr, struct fsdriver_data * data, size_t bytes,
 		 * return a partial result.  Otherwise return the error.
 		 */
 		if (r < 0)
-			return (off > 0) ? off : r;
+			return (off > 0) ? (ssize_t)off : r;
 
 		off += len;
 		pos += len;
@@ -151,7 +151,7 @@ fs_write(ino_t ino_nr, struct fsdriver_data * data, size_t bytes, off_t pos,
 		 * return a partial result.  Otherwise return the error.
 		 */
 		if (r < 0)
-			return (off > 0) ? off : r;
+			return (off > 0) ? (ssize_t)off : r;
 
 		off += r;
 		pos += r;
